modified graph creators structure
