+++ shaketune/commands/__init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
-from .axes_input_shaper import axes_shaper_calibration as axes_shaper_calibration
-from .axes_map import axes_map_calibration as axes_map_calibration
-from .belts_comparison import compare_belts_responses as compare_belts_responses
-from .static_freq import excitate_axis_at_freq as excitate_axis_at_freq
-from .vibrations_profile import create_vibrations_profile as create_vibrations_profile
+from .axes_map_calibration import axes_map_calibration as axes_map_calibration
+from .axes_shaper_calibration import axes_shaper_calibration as axes_shaper_calibration
+from .compare_belts_responses import compare_belts_responses as compare_belts_responses
+from .create_vibrations_profile import create_vibrations_profile as create_vibrations_profile
+from .excitate_axis_at_freq import excitate_axis_at_freq as excitate_axis_at_freq
