+++ shaketune/graph_creators/__init__.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
 
-from .graph_creator import AxesMapFinder as AxesMapFinder
-from .graph_creator import BeltsGraphCreator as BeltsGraphCreator
+from .axes_map_graph_creator import AxesMapGraphCreator as AxesMapGraphCreator
+from .belts_graph_creator import BeltsGraphCreator as BeltsGraphCreator
 from .graph_creator import GraphCreator as GraphCreator
-from .graph_creator import ShaperGraphCreator as ShaperGraphCreator
-from .graph_creator import StaticGraphCreator as StaticGraphCreator
-from .graph_creator import VibrationsGraphCreator as VibrationsGraphCreator
+from .shaper_graph_creator import ShaperGraphCreator as ShaperGraphCreator
+from .static_graph_creator import StaticGraphCreator as StaticGraphCreator
+from .vibrations_graph_creator import VibrationsGraphCreator as VibrationsGraphCreator
