+++ shaketune/graph_creators/static_graph_creator.py
@@ -17,6 +17,8 @@ from ..helpers.common_func import (
     parse_log,
 )
 from ..helpers.console_output import ConsoleOutput
+from ..shaketune_config import ShakeTuneConfig
+from .graph_creator import GraphCreator
 
 PEAKS_DETECTION_THRESHOLD = 0.05
 PEAKS_EFFECT_THRESHOLD = 0.12
@@ -30,6 +32,52 @@ KLIPPAIN_COLORS = {
     'dark_orange': '#F24130',
     'red_pink': '#F2055C',
 }
+
+
+class StaticGraphCreator(GraphCreator):
+    def __init__(self, config: ShakeTuneConfig):
+        super().__init__(config, 'static frequency')
+        self._freq = None
+        self._duration = None
+        self._accel_per_hz = None
+
+    def configure(self, freq: float, duration: float, accel_per_hz: float = None) -> None:
+        self._freq = freq
+        self._duration = duration
+        self._accel_per_hz = accel_per_hz
+
+    def create_graph(self) -> None:
+        if not self._freq or not self._duration or not self._accel_per_hz:
+            raise ValueError('freq, duration and accel_per_hz must be set to create the static frequency graph!')
+
+        lognames = self._move_and_prepare_files(
+            glob_pattern='shaketune-staticfreq_*.csv',
+            min_files_required=1,
+            custom_name_func=lambda f: f.stem.split('_')[1].upper(),
+        )
+        fig = static_frequency_tool(
+            lognames=[str(path) for path in lognames],
+            klipperdir=str(self._config.klipper_folder),
+            freq=self._freq,
+            duration=self._duration,
+            max_freq=200.0,
+            accel_per_hz=self._accel_per_hz,
+            st_version=self._version,
+        )
+        self._save_figure_and_cleanup(fig, lognames, lognames[0].stem.split('_')[-1])
+
+    def clean_old_files(self, keep_results: int = 3) -> None:
+        # Get all PNG files in the directory as a list of Path objects
+        files = sorted(self._folder.glob('*.png'), key=lambda f: f.stat().st_mtime, reverse=True)
+
+        if len(files) <= keep_results:
+            return  # No need to delete any files
+
+        # Delete the older files
+        for old_file in files[keep_results:]:
+            csv_file = old_file.with_suffix('.csv')
+            csv_file.unlink(missing_ok=True)
+            old_file.unlink()
 
 
 ######################################################################
