+++ shaketune/shaketune_config.py
@@ -8,11 +8,11 @@ KLIPPER_FOLDER = Path.home() / 'klipper'
 KLIPPER_LOG_FOLDER = Path.home() / 'printer_data/logs'
 RESULTS_BASE_FOLDER = Path.home() / 'printer_data/config/K-ShakeTune_results'
 RESULTS_SUBFOLDERS = {
-    'axesmap': 'axes_map',
-    'belts': 'belts',
-    'shaper': 'input_shaper',
-    'vibrations': 'vibrations',
-    'staticfreq': 'static_freq',
+    'axes map calibration': 'axes_map',
+    'belts comparison': 'belts',
+    'input shaper': 'input_shaper',
+    'vibrations profile': 'vibrations',
+    'static frequency': 'static_freq',
 }
